⚙️wip: renamed component ComandCenter to RoverCommandTerminal and improved rover control interface
