+++ src/components/rover/RoverCommandTerminal.vue
